# frontend/src/pages/API/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Swagger Game Service - OpenAPI 3.0",
    "description": "This is a sample Game Service Server based on the OpenAPI 3.0 specification.",
    "contact": {
      "email": "j.falkiewicz.089@studms.ug.edu.pl"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.11"
  },
  "paths": {
...
    "/score": {
      "post": {
        "tags": ["Score"],
        "summary": "Create a new score",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "number"
                  },
                  "private": {
                    "type": "boolean"
                  },
                  "game": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "player": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Score created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "object",
                      "properties": {
                        "msgBody": {
                          "type": "string"
                        },
                        "msgError": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error creating the score",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "object",
                      "properties": {
                        "msgBody": {
                          "type": "string"
                        },
                        "msgError": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Score"],
        "summary": "Get all scores",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Score"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error fetching scores",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "object",
                      "properties": {
                        "msgBody": {
                          "type": "string"
                        },
                        "msgError": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/score/scoreboard/{gameId}": {
      "get": {
        "tags": ["Score"],
        "summary": "Get scoreboard by game ID",
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "description": "Game ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Score"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error fetching scores",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "object",
                      "properties": {
                        "msgBody": {
                          "type": "string"
                        },
                        "msgError": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/score/{gameId}": {
      "get": {
        "tags": ["Score"],
        "summary": "Get scores by game ID",
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "description": "Game ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Score"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error fetching scores",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "object",
                      "properties": {
                        "msgBody": {
                          "type": "string"
                        },
                        "msgError": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Score": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "string"
          },
          "private": {
            "type": "boolean"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "game": {
            "type": "string",
            "format": "uuid"
          },
          "player": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": ["amount", "private", "game", "player"]
      },
...
    }
  }
}